+++ src/common/modules/browser/isFirefox.ts
