--- src/common/modules/getImageURL.ts
@@ -1,9 +1,0 @@
-import { isDevelopment } from '@common/modules/isDevelopment'
-
-export default (uri: string): string => {
-    if (isDevelopment()) {
-        return `${uri}`
-    }
-
-    return chrome.runtime.getURL(`${uri}`)
-}
