+++ src/common/modules/storage/deleteFromStorage.ts
@@ -1,5 +1,5 @@
 import { isDevelopment } from '@common/modules/isDevelopment'
-import { isFirefox } from '@common/modules/isFirefox'
+import { isFirefox } from '@common/modules/browser/isFirefox'
 
 export default async (key: string): Promise<void> => {
     if (isDevelopment()) {
