+++ src/common/modules/storage/getFromStorage.ts
@@ -1,5 +1,5 @@
 import { isDevelopment } from '@common/modules/isDevelopment'
-import { isFirefox } from '@common/modules/isFirefox'
+import { isFirefox } from '@common/modules/browser/isFirefox'
 import error from '@common/modules/error'
 
 export default <T>(key: string): Promise<T | null> => {
