+++ src/common/modules/trans.ts
@@ -1,5 +1,5 @@
 import { isDevelopment } from '@common/modules/isDevelopment'
-import { isFirefox } from '@common/modules/isFirefox'
+import { isFirefox } from '@common/modules/browser/isFirefox'
 
 let englishMessages: { [key: string]: { message: string } } | null = null
