+++ src/popup/components/Navbar/Sidebar/Heading.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { ref, onMounted, computed } from 'vue'
 import trans from '@common/modules/trans'
-import getImageURL from '@common/modules/getImageURL'
+import { getImageURL } from '@common/modules/browser/runtime'
 import { isDevelopment } from '@common/modules/isDevelopment'
-import { isFirefox } from '@common/modules/isFirefox'
+import { isFirefox } from '@common/modules/browser/isFirefox'
 import getLocalStorageUsage from '@common/modules/storage/getLocalStorageUsage'
 
 onMounted(setCurrentBytesUsage)
