+++ src/popup/components/Navbar/Sidebar/Sidebar.vue
@@ -2,8 +2,8 @@
 import { useSidebarStore } from '@/stores/sidebar'
 import trans from '@common/modules/trans'
 import openSettingsPage from '@/modules/openSettingsPage'
-import { isFirefox } from '@common/modules/isFirefox'
-import getImageURL from '@common/modules/getImageURL'
+import { isFirefox } from '@common/modules/browser/isFirefox'
+import { getImageURL } from '@common/modules/browser/runtime'
 import Hamburger from '@/components/Navbar/Sidebar/Hamburger.vue'
 import RightSlideTransition from '@common/components/Transitions/RightSlideTransition.vue'
 import Overlay from '@/components/Navbar/Sidebar/Overlay.vue'
